set Zero txt and github links in comments and docs

+++ packages/contracts/contracts/Interfaces/IPriceFeed.sol
@@ -4,14 +4,14 @@ pragma solidity 0.6.11;
 
 interface IPriceFeed {
     // --- Events ---
-    event LastGoodPriceUpdated(uint _lastGoodPrice);
+    event LastGoodPriceUpdated(uint256 _lastGoodPrice);
 
     // --- Function ---
 
-    /// @notice Returns the latest price obtained from the Oracle. Called by Liquity functions that require a current price.
+    /// @notice Returns the latest price obtained from the Oracle. Called by Zero functions that require a current price.
     ///         It uses the main price feed and fallback to the backup one in case of an error. If both fail return the last
     ///         good price seen.
     /// @dev It's also callable by anyone externally
     /// @return The price
-    function fetchPrice() external returns (uint);
+    function fetchPrice() external returns (uint256);
 }
